requirements.txt
# Core RAG
openai>=1.0.0  # OpenRouter uses OpenAI-compatible API
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-text-splitters>=0.1.0
chromadb>=0.4.0
pypdf2>=3.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
tiktoken>=0.5.0
sentence-transformers>=2.2.0
transformers>=4.36.0
rank-bm25>=0.2.2  # For hybrid search

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6

# Utilities
requests>=2.31.0
numpy>=1.24.0
pandas>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
aiofiles>=23.0.0

# UI
streamlit>=1.28.0

# Firebase
firebase-admin>=6.0.0

# MLOps & Monitoring
prometheus-client>=0.19.0
redis>=5.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
httpx>=0.25.0  # For testing FastAPI
